Add pagination and remote source for quotes.html
I really would like to do testing on this, but I'm less familiar with getting testing set up in JS and don't want to waste time.

I also would have liked to set up a router to get `pushState` for pagination - but it adds complexity and I didn't write any tests as it is.